refactor(ui): move ProgressBar component from renderer to packages/ui
Move ProgressBar to the UI library so it can be imported via
@podman-desktop/ui-svelte instead of cross-package relative
paths. Update consumer imports in TaskIndicator and
TaskManagerTableProgressColumnRunning.
